solution/requirements.txt
flask
requests
zowe-python-sdk-bundle
python-dotenv
